conversation_clustering/requirements.txt
dspy==3.1.0
mlflow==3.8.1
numpy==1.26.4
webvtt-py
datasets
